final push

+++ controllers/medicineController.js
@@ -1,5 +1,4 @@
 const medicineModel = require('../models/medicineModel');
-const userModel = require('../models/userModel.js')
 const validation = require('../validator/validation');
 const moment = require('moment');
 const fs = require('fs');
@@ -9,7 +8,6 @@ const uploadFileOnCloudinary = require('../imageUpload/cloudinary.js')
 const addMedicine = async function (req, res) {
     try {
         const data = req.body;
-        // Check if request body is empty 
         if (!validation.isEmpty(data)) {
             return res.status(400).send({ status: false, message: "Provide details to add medicine" });
         }
@@ -26,19 +24,16 @@ const addMedicine = async function (req, res) {
             return res.status(500).send({ status: false, message: "Failed to upload image to Cloudinary" });
         }
 
-        // Extract sellerId from accessToken 
         const sellerId = req.decodedToken.userId;
         const { category, medicineName, description, form, stockQuantity, price, currencyId,
             currencyFormat, expiryDate } = data;
 
-        // Validate mandatory details
         if (!validation.checkData(category)) {
             return res.status(400).send({ status: false, message: "Category is required" })
         }
         if (!validation.checkData(medicineName)) {
             return res.status(400).send({ status: false, message: "Medicine name is required" });
         }
-        // Check is provided medicine name already exist
         const isexistMedicine = await medicineModel.findOne({ seller:sellerId, medicineName: medicineName });
         if (isexistMedicine) {
             return res.status(409).send({ status: false, message: "Provided medicine name already exist" })
@@ -65,13 +60,13 @@ const addMedicine = async function (req, res) {
             return res.status(400).send({ status: false, message: "CurrencyId is required" });
         }
         if (!(/INR/.test(currencyId))) {
-            return res.status(400).send({ status: false, message: " CurrencyId should be in 'INR' Format" });
+            return res.status(400).send({ status: false, message: "CurrencyId should be in 'INR' Format" });
         }
         if (!validation.checkData(currencyFormat)) {
             return res.status(400).send({ status: false, message: "CurrencyFormat is required" });
         }
         if (!(/₹/.test(currencyFormat))) {
-            return res.status(400).send({ status: false, message: "Currency format of product should be in '₹' " });
+            return res.status(400).send({ status: false, message: "Currency format of medicine should be in '₹' " });
         }
 
         // Parsing expiry date of medicine using moment.js
@@ -103,7 +98,8 @@ const addMedicine = async function (req, res) {
         return res.status(500).send({ status: false, message: error.message });
     } finally {
         // Clean up the local image of medicine from your local system after processing
-        if (req.file && fs.existsSync(req.file.path)) {
+        if (req.file && req.file.path && fs.existsSync(req.file.path)) {
+            console.log("successfully removed file from local system")
             fs.unlinkSync(req.file.path);
         }
     }
@@ -120,9 +116,9 @@ const getMedicine = async function (req, res) {
         } else {
             // Filter provided, fetch medicines based on filter parameters
             let query = { isDeleted: false };
-            const { name, category, priceGreaterThan, priceLessThan } = filter;
-            if (name) {
-                query.medicineName = name;
+            const { medicineName, category, priceGreaterThan, priceLessThan } = filter;
+            if (medicineName) {
+                query.medicineName = medicineName;
             }
             if (category) {
                 query.category = category;
@@ -130,18 +126,18 @@ const getMedicine = async function (req, res) {
             if (priceGreaterThan && priceLessThan) {
                 query.price = { $gt: Number(priceGreaterThan), $lt: Number(priceLessThan) }
             }
-            if (priceGreaterThan) {
+           else if (priceGreaterThan) {
 
                 query.price = { $gt: Number(priceGreaterThan) }
             }
-            if (priceLessThan) {
+            else if(priceLessThan) {
                 query.price = { $lt: Number(priceLessThan) }
             }
 
             getData = await medicineModel.find(query)
         }
 
-        return res.status(200).send({ status: false, message: "Fetched detail successfully", data: getData });
+        return res.status(200).send({ status: true, message: "Fetched detail successfully", data: getData });
 
     } catch (error) {
         return res.status(500).send({ status: false, message: error.message });
@@ -153,55 +149,43 @@ const updateMedicine = async function (req, res) {
     try {
         const medicineID = req.params.medicineId;
         if (!validation.checkObjectId(medicineID)) {
-            return res.status(400).send({ status: false, message: " Invalid objectId " })
+            return res.status(400).send({ status: false, message: "Invalid objectId" })
         }
 
-        const isexistMedicine = await medicineModel.findById({ _id: medicineID });
+        const isexistMedicine = await medicineModel.findById(medicineID );
         if (!isexistMedicine) {
-            return res.status(400).send({ status: false, message: " Provided medicine objectId does not exist " })
+            return res.status(400).send({ status: false, message: "Medicine not found" })
         }
 
         const sellerId = isexistMedicine.seller;
         // Check authorization: Only the seller who listed the medicine can update it
-        if (req.decodedToken.userId != sellerId) {
-            return res.status(403).send({ status: false, message: " Unauthorized to update " })
+        if (req.decodedToken.userId !== sellerId) {
+            return res.status(403).send({ status: false, message: "Unauthorized to update" })
         }
 
         // Only not deleted medicine can be updated 
         let updatedField = { isDeleted: false }
 
-        // Check if seller updates medicine image
+         // Handle medicine image update
         if (req.file) {
             // multer uploaded file inside req.file property
             const medicineImage = req.file.path;
             // Upload file in cloudinary
             const cloudinaryResponse = await uploadFileOnCloudinary(medicineImage);
             if (!cloudinaryResponse) {
-                return res.status(500).send({ status: false, message: " Failed to upload image to Cloudinary " });
+                return res.status(500).send({ status: false, message: "Failed to upload image to Cloudinary" });
             }
             updatedField.medicineImage = cloudinaryResponse.url;
         }
 
-        // Check if request body is empty 
-        if (!validation.isEmpty(req.body)) {
-            return res.status(400).send({ status: false, message: "Provide details for update" });
-        }
+        // Handle other fields to update
         const { category, medicineName, description, form, stockQuantity, price, currencyId,
             currencyFormat, expiryDate } = req.body;
-        // Validate updated fields
+            
         if (category) {
-            const isexistCategory = await medicineModel.findOne({ category: category })
-            if (isexistCategory) {
-                return res.status(400).send({ status: false, message: " Cannot Update, category is already exist " })
-            }
             updatedField.category = category;
         }
         if (medicineName) {
-            //const isexistMedicine = await medicineModel.findOne({ medicineName: medicineName })
-            const nameofMedicine = isexistMedicine.medicineName;
-            if (medicineName === nameofMedicine) {
-                return res.status(400).send({ status: false, message: " Cannot Update, medicine name is already exist " })
-            }
             updatedField.medicineName = medicineName;
         }
         if (description) {
@@ -209,7 +193,7 @@ const updateMedicine = async function (req, res) {
         }
         if (form) {
             if (!["tablet", "capsule", "syrup"].includes(form)) {
-                return res.status(400).send({ status: false, message: " Form only include tablet,capsule and syrup " })
+                return res.status(400).send({ status: false, message: "Form only include tablet,capsule and syrup" })
             }
             updatedField.form = form;
         }
@@ -218,14 +202,14 @@ const updateMedicine = async function (req, res) {
         }
         if (price) {
             if (!validation.isValidPrice(price)) {
-                return res.status(400).send({ status: false, message: " Enter a valid price" });
+                return res.status(400).send({ status: false, message: "Enter a valid price" });
             }
             updatedField.price = price;
         }
         if (currencyId) {
 
             if (!(/INR/.test(currencyId))) {
-                return res.status(400).send({ status: false, message: " CurrencyId should be in 'INR' Format" });
+                return res.status(400).send({ status: false, message: "CurrencyId should be in 'INR' Format" });
             }
 
             updatedField.currencyId = currencyId;
@@ -233,7 +217,7 @@ const updateMedicine = async function (req, res) {
         if (currencyFormat) {
 
             if (!(/₹/.test(currencyFormat))) {
-                return res.status(400).send({ status: false, message: " Currency format of medicine should be in '₹' " });
+                return res.status(400).send({ status: false, message: "Currency format of medicine should be in '₹' "});
             }
 
             updatedField.currencyFormat = currencyFormat
@@ -242,19 +226,20 @@ const updateMedicine = async function (req, res) {
             // Parsing expiry date of medicine using moment.js
             const expiredDateofMedicine = moment(expiryDate, 'YYYY-MM-DD');
             if (!expiredDateofMedicine.isValid()) {
-                return res.status(400).send({ status: false, message: " Invalid date format " });
+                return res.status(400).send({ status: false, message: "Invalid date format" });
             }
             updatedField.expiryDate = expiryDate;
         }
 
         const medicineUpdate = await medicineModel.findByIdAndUpdate({ _id: medicineID }, updatedField, { new: true });
-        return res.status(200).send({ status: true, message: " Updated Successfully ", data: medicineUpdate });
+        return res.status(200).send({ status: true, message: "Updated Successfully", data: medicineUpdate });
 
     } catch (error) {
         return res.status(500).send({ status: false, message: error.message });
     } finally {
         // Clean up the local image of medicine from your local system after processing
-        if (req.file && fs.existsSync(req.file.path)) {
+        if (req.file && req.file.path && fs.existsSync(req.file.path)) {
+            console.log("successfully removed file from local system")
             fs.unlinkSync(req.file.path);
         }
     }
@@ -265,31 +250,30 @@ const deleteMedicine = async function (req, res) {
     try {
         const medicineID = req.params.medicineId;
         if (!validation.checkObjectId(medicineID)) {
-            return res.status(400).send({ status: false, message: " Invalid objectId " })
+            return res.status(400).send({ status: false, message: "Invalid objectId" })
         }
 
-        const isexistMedicine = await medicineModel.findById({ _id: medicineID });
+        const isexistMedicine = await medicineModel.findById(medicineID );
         if (!isexistMedicine) {
-            return res.status(400).send({ status: false, message: " Provided medicine objectId does not exist " })
+            return res.status(400).send({ status: false, message: "Provided medicine objectId does not exist" })
         }
 
         const sellerId = isexistMedicine.seller;
         // Check authorization: Only the seller who listed the medicine can delete it
         if (req.decodedToken.userId != sellerId) {
-            return res.status(403).send({ status: false, message: " Unauthorized to update " })
+            return res.status(403).send({ status: false, message: "Unauthorized to update" })
         }
 
-        if (isexistMedicine.isDeleted === false) {
-            await medicineModel.findByIdAndUpdate({ _id: medicineID }, { $set: { isDeleted: true } })
+        if(isexistMedicine.isDeleted === true){
+            return res.status(400).send({status:false, message: "Medicine is already deleted"})
         }
 
-        return res.status(200).send({ status: true, message: "Deleted Successfully" })
-
+        await medicineModel.findByIdAndUpdate({ _id: medicineID }, { $set: { isDeleted: true } })
+        return res.status(200).send({ status: true, message: "Medicine deleted successfully" })
     } catch (error) {
         return res.status(500).send({ status: false, message: error.message });
     }
 }
-
 
 
 module.exports = { addMedicine, getMedicine, updateMedicine, deleteMedicine };
